# servicios/nucleo-financiero/src/main/resources/openapi/nucleo-financiero.yaml
...
openapi: 3.1.0
info:
  title: nucleo-financiero
  version: 1.0.0
  description: >-
    Servicio nucleo-financiero. Un caso de uso = una operacion, con el codigo CU en el
    operationId. Los importes viajan como CADENA decimal, nunca como number.
servers:
  - url: /api/v1
paths:
  /billetera/{cuentaId}/saldo:
    get:
      operationId: consultarSaldo
      summary: Saldo de una cuenta de billetera
      description: >-
        El saldo NO se guarda: se deriva del libro. Esta operacion existe en el
        borrador porque los trece servicios necesitan saber que la forma del saldo
        es Dinero, no un number.
      parameters:
        - name: cuentaId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Saldo disponible y retenido
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SaldoBilletera' }
        '401':
          $ref: '#/components/responses/NoAutenticado'
  /billetera/cotizaciones:
    post:
      operationId: cotizarOperacion            # CU-30 (consume tarifas)
      summary: Ver el precio de una recarga o un retiro antes de confirmarlo
      description: >-
        [CONTRATO] Que la persona vea el numero final ANTES de aceptar: base, comision, impuesto,
        costo total, neto y hasta cuando vale ese precio. No mueve nada ni compromete a nadie. La
        `Idempotency-Key` que se manda aca es la MISMA que despues lleva la recarga o el retiro, y
        `cotizacionId` es lo que esa operacion cita para confirmar. Pedir otro importe con la misma
        clave produce otra cotizacion: no se puede cobrar el precio de un importe sobre otro.

        El precio lo fija `tarifas`; este servicio no lo calcula. Una operacion sin concepto en el
        tarifario es gratuita (`gratuita=true`, sin `cotizacionId`).
      parameters:
        - $ref: '#/components/parameters/ClaveIdempotencia'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EntradaCotizacionOperacion' }
      responses:
        '200':
          description: La cotizacion y su desglose
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SalidaCotizacionOperacion' }
        '422':
          $ref: '#/components/responses/ReglaDeNegocio'
          # AP-CU10-05 / AP-CU11-01 no se pudo cotizar (tarifas no respondio)
          # AP-CU10-08 / AP-CU11-12 el costo no puede igualar ni superar el importe
        '401':
          $ref: '#/components/responses/NoAutenticado'
  /billetera/recargas:
    post:
      operationId: solicitarRecarga           # CU-10
      summary: Abre una orden de recarga y devuelve cuanto se va a acreditar
      description: >-
        [CONTRATO] Derivada de la implementacion del carril 2A, que llego antes que
        su contrato.

        La orden NO acredita: reserva la intencion y vence. Acreditar es el paso
        siguiente y ocurre cuando el proveedor confirma, porque acreditar contra una
        promesa es acreditar plata que puede no llegar.

        `acreditara` viaja en la respuesta ya calculado: el cliente **no** resta el
        costo del proveedor por su cuenta. Un cliente que recalcula produce un numero
        que no es el del libro contable.
      parameters:
        - $ref: '#/components/parameters/ClaveIdempotencia'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EntradaRecarga' }
      responses:
        '201':
          description: Orden abierta, con su vencimiento
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SalidaRecarga' }
        '422':
          $ref: '#/components/responses/ReglaDeNegocio'
          # AP-CU10-01 LIMITE_EXCEDIDO           · AP-CU10-02 SALDO_MAXIMO_ALCANZADO
          # AP-CU10-03 INSTRUMENTO_NO_VERIFICADO · AP-CU10-04 CUENTA_NO_OPERATIVA
          # AP-CU10-05 ORDEN_EXPIRADA
          # AP-CU10-06 COTIZACION_REQUERIDA · AP-CU10-07 COTIZACION_VENCIDA
          # AP-CU10-08 COTIZACION_NO_COINCIDE
        '401':
          $ref: '#/components/responses/NoAutenticado'
  /billetera/recargas/{ordenId}/acreditacion:
    post:
      operationId: acreditarRecarga           # CU-10
      summary: Acredita una orden que el proveedor confirmo
      description: >-
        [CONTRATO] La confirmacion del proveedor. Es idempotente por la orden: si dos
        confirmaciones llegan juntas, una acredita y la otra encuentra la orden ya
        acreditada — la red duplica, y eso no puede duplicar el saldo.
      parameters:
        - name: ordenId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/ClaveIdempotencia'
      responses:
        '200':
          description: Acreditada, con el saldo resultante
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SalidaAcreditacion' }
        '422':
          $ref: '#/components/responses/ReglaDeNegocio'
          # AP-CU10-05 ORDEN_EXPIRADA
        '401':
          $ref: '#/components/responses/NoAutenticado'
  /billetera/retenciones:
    post:
      operationId: retenerSaldo               # CU-13
      summary: Aparta saldo sin sacarlo de la cuenta
      description: >-
        [CONTRATO] Retener no es debitar: el dinero sigue siendo del titular y por eso
        se ve como retenido y no desaparece. Lo que cambia es que deja de estar
        disponible mientras dure el motivo.
      parameters:
        - $ref: '#/components/parameters/ClaveIdempotencia'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EntradaRetencion' }
      responses:
        '201':
          description: Retenido, con el saldo disponible y el retenido despues
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SalidaRetencion' }
        '422':
          $ref: '#/components/responses/ReglaDeNegocio'
          # AP-CU13-01 SALDO_INSUFICIENTE · AP-CU13-02 VENCIMIENTO_REQUERIDO
          # AP-CU13-03 RETENCION_YA_EJECUTADA
        '401':
          $ref: '#/components/responses/NoAutenticado'
  /billetera/retenciones/{retencionId}/cierre:
    post:
      operationId: cerrarRetencion            # CU-13
      summary: Libera o ejecuta una retencion
      description: >-
        [CONTRATO] Una retencion se cierra una sola vez. Si otra operacion la cerro
        primero, la segunda no la vuelve a cerrar: se responde
        `RETENCION_YA_EJECUTADA`, que no es una falla sino concurrencia sana.
      parameters:
        - name: retencionId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/ClaveIdempotencia'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EntradaCierreRetencion' }
      responses:
        '200':
          description: Cerrada, con el saldo resultante
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SalidaCierreRetencion' }
        '422':
          $ref: '#/components/responses/ReglaDeNegocio'
          # AP-CU13-03 RETENCION_YA_EJECUTADA
        '401':
          $ref: '#/components/responses/NoAutenticado'

  # CU-40 Evaluar limites NO tiene endpoint propio: corre DENTRO de la transaccion
  # de la operacion que evalua. Exponerlo dejaria que alguien consulte el limite,
  # vea que entra, y opere despues contra un acumulado ya cambiado.
  #   AP-CU40-01 SIN_LIMITE_CONFIGURADO · AP-CU40-02 LIMITE_EXCEDIDO
  #   AP-CU40-03 CANTIDAD_EXCEDIDA
  #
  # CU-57 Operar un punto de atencion quedo OBSOLETO con su numero reservado:
  # ADR-039 retiro el efectivo del alcance el 20-08-2026.
  /billetera/retiros:
    post:
      operationId: solicitarRetiro             # CU-11
      summary: Solicitar un retiro hacia una cuenta de destino
      description: >-
        Retencion primero, pago despues, nunca al reves. Si se instruyera el pago
        antes de retener, entre las dos cosas la persona podria gastar el mismo
        saldo en otra operacion y el retiro saldria contra un disponible que ya no
        existe. El proveedor se instruye FUERA de la transaccion (invariante 6).
      parameters:
        - $ref: '#/components/parameters/ClaveIdempotencia'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EntradaRetiro' }
      responses:
        '201':
          description: Orden creada y saldo ya retenido
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SalidaRetiro' }
        '422':
          $ref: '#/components/responses/ReglaDeNegocio'
          # AP-CU11-01 SALDO_INSUFICIENTE          · AP-CU11-02 MFA_REQUERIDO
          # AP-CU11-03 INSTRUMENTO_EN_ENFRIAMIENTO · AP-CU11-04 TITULAR_NO_COINCIDE
          # AP-CU11-05 BLOQUEO_DE_AUTORIDAD        · AP-CU11-06 ENCAJE_INCUMPLIDO
          # AP-CU11-07 MFA_INVALIDO (H2: evidencia provista pero no valida)
          # AP-CU11-13 la Idempotency-Key ya identifica otro retiro (otro importe, destino o cotizacion)
        '401':
          $ref: '#/components/responses/NoAutenticado'
  /billetera/retiros/{ordenId}/aprobacion:
    post:
      operationId: aprobarRetiro                # CU-11 (H3.S2, ADR-049)
      summary: Aprueba o rechaza un retiro que necesita doble aprobacion
      description: >-
        Solo alguien con RETIRO_APROBAR, DISTINTO de quien solicito el retiro
        (R-SEG-04): la segregacion la exige la aplicacion Y la base
        (ck_retiro_doble_aprobacion), dos capas independientes. Dos aprobadores
        simultaneos producen una sola transicion: el segundo recibe 409, nunca
        un doble efecto.
      parameters:
        - name: ordenId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/ClaveIdempotencia'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EntradaAprobacionRetiro' }
      responses:
        '200':
          description: Orden autorizada o rechazada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SalidaAprobacionRetiro' }
        '422':
          $ref: '#/components/responses/ReglaDeNegocio'
          # AP-CU11-10 RETIRO_AUTOAPROBACION_PROHIBIDA
        '404':
          description: Esa orden no existe
        '409':
          description: >-
            La orden ya no esta en revision — otro aprobador la resolvio primero, o
            no requiere doble aprobacion.
        '403':
          description: Sin el permiso RETIRO_APROBAR
        '401':
          $ref: '#/components/responses/NoAutenticado'
  /billetera/transferencias:
    post:
      operationId: transferirSaldo             # CU-12
      summary: Transferir saldo entre billeteras
      description: >-
        La unica operacion donde el dinero no entra ni sale del sistema: cambia de
        bolsillo, y el saldo total permanece constante. Las dos cuentas se bloquean
        en orden de identificador — dos transferencias cruzadas que bloqueen cada
        una primero la suya se esperan para siempre.
      parameters:
        - $ref: '#/components/parameters/ClaveIdempotencia'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EntradaTransferencia' }
      responses:
        '201':
          description: Transferencia ejecutada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SalidaTransferencia' }
        '422':
          $ref: '#/components/responses/ReglaDeNegocio'
          # AP-CU12-01 SALDO_INSUFICIENTE          · AP-CU12-02 DESTINO_INEXISTENTE
          # AP-CU12-03 CUENTA_DESTINO_NO_OPERATIVA · AP-CU12-04 LIMITE_EXCEDIDO
          # AP-CU12-05 P2P_NO_HABILITADO
        '401':
          $ref: '#/components/responses/NoAutenticado'
  /billetera/reversos:
    post:
      operationId: reversarTransaccion         # CU-14
      summary: Reversar una transaccion aplicada
      description: >-
        Reversar es escribir el espejo, no borrar el original. El libro es
        append-only y la transaccion original conserva sus movimientos sin un solo
        cambio: el extracto tiene que poder contar la historia completa —el cobro y
        su devolucion—, no un cobro que se desvanecio.
        Cuatro ojos: quien autoriza no puede ser quien solicita.
      parameters:
        - $ref: '#/components/parameters/ClaveIdempotencia'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EntradaReverso' }
      responses:
        '201':
          description: Reverso ejecutado, o restitucion generada si el saldo no alcanza
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SalidaReverso' }
        '422':
          $ref: '#/components/responses/ReglaDeNegocio'
          # AP-CU14-01 TRANSACCION_NO_REVERSABLE · AP-CU14-02 FALTA_AUTORIZACION
          # AP-CU14-03 SALDO_INSUFICIENTE_PARA_REVERSO
        '401':
          $ref: '#/components/responses/NoAutenticado'
  /billetera/{cuentaId}/extracto:
    get:
      operationId: emitirExtracto              # CU-15
      summary: Emitir el extracto de un periodo
      description: >-
        Si el extracto no cuadra con el cierre diario, NO se emite: uno con una
        cifra que no coincide es peor que no tenerlo — la persona lo usa para un
        tramite, se lo rechazan, y descubre el error cuando ya le costo algo. El
        descuadre queda registrado igual, que es lo que pide el criterio.
      parameters:
        - name: cuentaId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: desde
          in: query
          required: true
          schema: { type: string, format: date }
        - name: hasta
          in: query
          required: true
          schema: { type: string, format: date }
      responses:
        '200':
          description: Extracto emitido, o bloqueado con su motivo
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SalidaExtracto' }
        '422':
          $ref: '#/components/responses/ReglaDeNegocio'
          # AP-CU15-01 PERIODO_SIN_CIERRES · AP-CU15-02 EXTRACTO_NO_CUADRA
          # AP-CU15-03 SIN_PERMISO_SOBRE_LA_CUENTA
        '401':
          $ref: '#/components/responses/NoAutenticado'
  /billetera/cierres:
    post:
      operationId: solicitarCierreBilletera    # CU-16
      summary: Solicitar el cierre de la billetera
      description: >-
        Cerrar NO borra: la cuenta queda CERRADA y sus movimientos siguen
        disponibles todo el plazo de conservacion. Cuatro puertas la frenan, y las
        cuatro protegen a terceros antes que a la empresa — un cierre que deje un
        aporte sin pagar traslada la perdida a los otros del pasanaku.
      parameters:
        - $ref: '#/components/parameters/ClaveIdempotencia'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EntradaCierre' }
      responses:
        '201':
          description: Solicitud registrada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SalidaCierreBilletera' }
        '422':
          $ref: '#/components/responses/ReglaDeNegocio'
          # AP-CU16-01 OBLIGACIONES_ABIERTAS · AP-CU16-02 GRUPO_ACTIVO
          # AP-CU16-03 RETENCION_VIGENTE     · AP-CU16-04 BLOQUEO_DE_AUTORIDAD
        '401':
          $ref: '#/components/responses/NoAutenticado'
  /billetera/bloqueos:
    post:
      operationId: bloquearPorAutoridad        # CU-17
      summary: Registrar un bloqueo por orden de autoridad
      description: >-
        El unico caso en que una retencion no tiene fecha de fin: la levanta la
        misma autoridad que la puso, y ponerle vencimiento propio seria decidir por
        un juez cuando se acaba su orden. El documento de respaldo es obligatorio —
        inmovilizar la plata de alguien sin poder mostrar el papel es indefendible.
      parameters:
        - $ref: '#/components/parameters/ClaveIdempotencia'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EntradaBloqueo' }
      responses:
        '201':
          description: Bloqueo vigente y saldo inmovilizado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SalidaBloqueo' }
        '422':
          $ref: '#/components/responses/ReglaDeNegocio'
          # AP-CU17-01 OFICIO_DUPLICADO · AP-CU17-02 CUENTA_INEXISTENTE
          # AP-CU17-03 SIN_DOCUMENTO_DE_RESPALDO
        '401':
          $ref: '#/components/responses/NoAutenticado'
  /billetera/qr:
    post:
      tags: [qr]
      operationId: emitirQr                    # CU-12 · QR interno
      summary: Emitir un QR interno para cobrar
      description: >-
        [CONTRATO] Un QR INTERNO es un cobro que una billetera de Pasanaku publica; se paga
        entre billeteras de Pasanaku. El DINAMICO fija importe y vencimiento y se consume al
        pagarse. El ESTATICO no fija importe, no vence y se puede pagar muchas veces.

        El `contenido` solo lleva un identificador y una firma: alterarlo lo invalida, nunca
        cambia a quien se paga ni cuanto. NO es un QR bancario ni interoperable: ese no se
        procesa (AP-CU12-10).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EntradaQr' }
      responses:
        '201':
          description: QR emitido
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SalidaQrEmitido' }
        '422':
          $ref: '#/components/responses/ReglaDeNegocio'
          # AP-CU12-02 solo el titular · AP-CU12-03 CUENTA_NO_OPERATIVA · AP-CU12-07 QR_NO_VALIDO
        '401':
          $ref: '#/components/responses/NoAutenticado'
...
components:
  parameters:
    ClaveIdempotencia:
      name: Idempotency-Key
      in: header
      required: true
      description: >-
        Toda operacion con efecto la exige. La red duplica; reintentar tiene que ser
        seguro, y la clave se valida ANTES de escribir (invariante 7).
      schema: { type: string, format: uuid }
  schemas:
...
    EntradaCotizacionOperacion:
      type: object
      additionalProperties: false
      required: [operacion, cuentaBilleteraId, monto]
      properties:
        operacion: { type: string, enum: [RECARGA, RETIRO] }
        cuentaBilleteraId: { type: string, format: uuid }
        monto: { $ref: '#/components/schemas/Dinero' }
    SalidaCotizacionOperacion:
      type: object
      additionalProperties: false
      required: [base, comision, impuesto, costoTotal, neto, gratuita]
      properties:
        cotizacionId:
          type: string
          format: uuid
          description: la version del precio que se acepta; ausente cuando la operacion es gratuita
        base: { $ref: '#/components/schemas/Dinero' }
        comision: { $ref: '#/components/schemas/Dinero' }
        impuesto: { $ref: '#/components/schemas/Dinero' }
        costoTotal: { $ref: '#/components/schemas/Dinero' }
        neto:
          $ref: '#/components/schemas/Dinero'
        validaHasta: { type: string, format: date-time }
        gratuita: { type: boolean }
...
    SalidaAprobacionRetiro:
      type: object
      additionalProperties: false
      required: [ordenRetiroId, estado, aprobadaPor]
      properties:
        ordenRetiroId: { type: string, format: uuid }
        estado:        { type: string, enum: [AUTORIZADA, RECHAZADA] }
        aprobadaPor:   { type: string, format: uuid }
    EntradaRecarga:
      type: object
      additionalProperties: false
      required: [cuentaBilleteraId, monto, medio]
      properties:
        cuentaBilleteraId: { type: string, format: uuid }
        monto: { $ref: '#/components/schemas/Dinero' }
        costoProveedor: { $ref: '#/components/schemas/Dinero' }
        medio: { type: string, maxLength: 30 }
        instrumentoFondeoId: { type: string, format: uuid }
        cotizacionId:
          type: string
          format: uuid
          description: >-
            La cotizacion que la persona vio y acepto (POST /billetera/cotizaciones, con la MISMA
            Idempotency-Key). Obligatoria mientras `aportaya.cotizacion.exigir-previa` este activa.
    SalidaRecarga:
      type: object
      additionalProperties: false
      required: [ordenRecargaId, estado, expiraEn, acreditara]
      properties:
        ordenRecargaId: { type: string, format: uuid }
        estado: { type: string, maxLength: 20 }
        expiraEn: { type: string, format: date-time }
        acreditara:
          $ref: '#/components/schemas/Dinero'
    SalidaAcreditacion:
      type: object
      additionalProperties: false
      required: [ordenRecargaId, transaccionId, saldoDespues]
      properties:
        ordenRecargaId: { type: string, format: uuid }
        transaccionId:  { type: string, format: uuid }
        saldoDespues:   { $ref: '#/components/schemas/Dinero' }
...
    SaldoBilletera:
      type: object
      additionalProperties: false
      required: [cuentaId, disponible, retenido, alCorteDe]
      properties:
        cuentaId:   { type: string, format: uuid }
        disponible: { $ref: '#/components/schemas/Dinero' }
        retenido:   { $ref: '#/components/schemas/Dinero' }
        alCorteDe:  { type: string, format: date-time }
  responses:
    ReglaDeNegocio:
      description: Regla de negocio de la aplicacion (422, no 400)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NoAutenticado:
      description: Sin sesion valida
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }